backend/requirements.txt
# Backend API Dependencies

# Core Web Framework
fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
pydantic>=2.5
pydantic-settings>=2.1

# Deep Learning
torch>=2.0
torchvision>=0.15
onnx>=1.15
onnxruntime>=1.16

# Computer Vision
opencv-python>=4.8
Pillow>=10.0
numpy>=1.24

# Face Recognition
facenet-pytorch>=2.5

# Vector Search
faiss-cpu>=1.7

# Database
sqlalchemy>=2.0
aiosqlite>=0.19

# Authentication & Security
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7

# Utilities
python-dotenv>=1.0
tqdm>=4.66
requests>=2.31
aiofiles>=23.0

# Logging
loguru>=0.7

# Testing
pytest>=7.4
pytest-asyncio>=0.21
httpx>=0.25

# Data Processing
pandas>=2.1
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.13
openpyxl>=3.1
